improve file structure again
improve the file structure again to remove the methods from the main file and require them into the logic file.
